Fixing issue with 0.01 on the screen; using type_t

+++ Panel.h
@@ -2,21 +2,21 @@
 #include "config.h"
 #include "TM1637Display.h"
 
-#define PANEL_BRIGHTNESS 4
-#define PANEL_SPLASH_DURATION 2000
+#define PANEL_BRIGHTNESS 4 // 0..7
+#define PANEL_SPLASH_DURATION 2000 // ms
 
 class Panel {
 
   private:
-    int8_t _pinCLK;
-    int8_t _pinDIO;
+    uint8_t _pinCLK;
+    uint8_t _pinDIO;
     TM1637Display _tm;
     
   public:
-    Panel(int8_t pinCLK, int8_t pinDIO);
-    void setup(int8_t mode);
-    void displayTime(float time);
+    Panel(uint8_t pinCLK, uint8_t pinDIO);
+    void setup(uint8_t mode);
+    void displayTime(uint32_t time);
     void displayFalseStart();
     void off();
-    void error(int8_t error);
+    void error(uint8_t error);
 };
